refactor: move data-value-mutations into shared folder
