+++ src/data-workspace/data-entry-cell/entry-field-input.js
@@ -3,9 +3,9 @@ import React, { useCallback, useMemo } from 'react'
 import { useSetRightHandPanel } from '../../right-hand-panel/index.js'
 import {
     VALUE_TYPES,
+    dataDetailsSidebarId,
     useSetHighlightedFieldIdContext,
 } from '../../shared/index.js'
-import { dataDetailsSidebarId } from '../constants.js'
 import { focusNext, focusPrev } from '../focus-utils/index.js'
 import {
     GenericInput,
