+++ src/data-workspace/data-entry-cell/inner-wrapper.js
@@ -5,13 +5,12 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { useField } from 'react-final-form'
 import {
+    mutationKeys as dataValueMutationKeys,
+    useDataValueParams,
     useHighlightedFieldIdContext,
     useValueStore,
 } from '../../shared/index.js'
-import {
-    useDataValueParams,
-    mutationKeys as dataValueMutationKeys,
-} from '../data-value-mutations/index.js'
+import { useHasComment } from '../has-comment/has-comment-context.js'
 import styles from './data-entry-cell.module.css'
 
 /** Three dots or triangle in top-right corner of cell */
