+++ src/data-workspace/index.js
@@ -1,5 +1,4 @@
 export * from './constants.js'
-export * from './validation/index.js'
 export { DataDetailsSidebar } from './data-details-sidebar/index.js'
 export { DataWorkspace } from './data-workspace.js'
 export * from './query-key-factory.js'
