+++ src/data-workspace/query-key-factory.js
@@ -42,30 +42,31 @@ export const dataValueContext = {
     ],
 }
 
-export const getValidationMetaDataQueryKey = (datasetId) => {
-    const queryKey = [
-        `validationRules`,
+export const getCompleteFormQueryKey = ({
+    dataSetId,
+    periodId,
+    orgUnitId,
+    categoryComboId,
+    categoryOptionIds,
+}) => {
+    const params = {
+        pe: periodId,
+        ou: orgUnitId,
+        // @TODO: Revisit these params to filter by AOC https://dhis2.atlassian.net/browse/TECH-1287
+        cc: categoryComboId,
+        cp: categoryOptionIds?.join(';'),
+    }
+
+    const validationQueryKey = [
+        `validation/dataSet/${dataSetId}`,
         {
-            params: {
-                dataSet: datasetId,
-                fields: [
-                    'id',
-                    'importance',
-                    'operator',
-                    'leftSide',
-                    'rightSide',
-                    'displayInstruction',
-                    'displayDescription',
-                    'displayName',
-                ],
-            },
+            params,
         },
     ]
-
-    return queryKey
+    return validationQueryKey
 }
 
-export const getValidationQueryKey = ({
+export const getIncompleteFormQueryKey = ({
     dataSetId,
     periodId,
     orgUnitId,
